@@ -2,14 +2,14 @@ apply plugin: 'com.android.application'
 apply plugin: 'io.fabric'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 28
+    buildToolsVersion '28.0.2'
     defaultConfig {
         applicationId "com.maths22.laundryview"
         minSdkVersion 16
-        targetSdkVersion 27
-        versionCode 9
-        versionName '1.0'
+        targetSdkVersion 28
+        versionCode 10
+        versionName '1.0.1'
         multiDexEnabled true
     }
     buildTypes {
@@ -26,28 +26,39 @@ android {
     }
     productFlavors {
     }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
 }
 
 dependencies {
-    annotationProcessor 'com.google.dagger:dagger-compiler:2.0.1'
-    annotationProcessor 'com.jakewharton:butterknife:7.0.1'
-    annotationProcessor 'com.google.guava:guava:22.0'
-    compileOnly 'com.google.dagger:dagger-compiler:2.0.1'
+    annotationProcessor 'com.google.dagger:dagger-compiler:2.17'
+    annotationProcessor 'com.jakewharton:butterknife-compiler:8.8.1'
+    annotationProcessor 'com.google.guava:guava:26.0-android'
     implementation fileTree(include: ['*.jar'], dir: 'libs')
     implementation 'com.android.support:multidex:1.0.3'
     implementation 'com.android.support:support-v13:27.1.1'
     implementation 'com.android.support:appcompat-v7:27.1.1'
     implementation 'com.android.support:design:27.1.1'
-    implementation 'com.google.dagger:dagger:2.0.1'
     implementation 'org.glassfish:javax.annotation:10.0-b28'
-    implementation 'com.jakewharton:butterknife:7.0.1'
+    implementation 'com.jakewharton:butterknife:8.8.1'
     implementation 'com.squareup.okhttp3:okhttp:3.9.0'
+    implementation 'com.google.dagger:dagger:2.17'
+    implementation 'com.google.guava:guava:26.0-android'
     implementation 'com.google.api-client:google-api-client-android:1.25.0'
-    implementation 'com.google.android.gms:play-services-analytics:16.0.3'
-    implementation 'com.google.firebase:firebase-messaging:17.3.2'
-    implementation 'com.google.firebase:firebase-core:16.0.3'
+    implementation 'com.google.android.gms:play-services-analytics:16.0.4'
+    implementation 'com.google.firebase:firebase-messaging:17.3.3'
+    implementation 'com.google.firebase:firebase-core:16.0.4'
     implementation 'com.crashlytics.sdk.android:crashlytics:2.9.5'
     testImplementation 'junit:junit:4.12'
+}
+
+configurations {
+    all {
+        exclude module: 'httpclient'
+        exclude module: 'commons-logging'
+    }
 }
 
 
