@@ -1,6 +1,7 @@
 package com.maths22.laundryview;
 
 import android.content.Context;
+import android.support.annotation.NonNull;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -20,8 +21,9 @@ public class SchoolArrayAdapter extends ArrayAdapter<School> {
         super(context, 0, users);
     }
 
+    @NonNull
     @Override
-    public View getView(int position, View convertView, ViewGroup parent) {
+    public View getView(int position, View convertView, @NonNull ViewGroup parent) {
         // Get the data item for this position
         School school = getItem(position);
         // Check if an existing view is being reused, otherwise inflate the view
